pyproject.toml
[project]
name = "mindbridge-chess"
version = "0.1.0"
description = "BCI chess using Unicorn Hybrid Black EEG and P300-based selection."
readme = "README.md"
requires-python = ">=3.11"
authors = [ { name = "MindBridge Team" } ]
dependencies = [
    "unicornpy>=0.1",
    "psychopy>=2023.2",
    "mne>=1.0",
    "python-chess>=2.0",
    "scikit-learn>=1.3",
    "numpy>=1.26",
    "pandas>=2.1"
]

[project.urls]
"Homepage" = "https://github.com/your-org/mindbridge-chess"
